\name{R/atlantis_consumption.R}
\alias{atlantis_stomach_matrix}
\alias{atlantis_stomach_content}

\title{Atlantis stomach content parsers}
\description{
    Routines to read and make use of the atlantis diet matrix
}

\usage{
atlantis_stomach_matrix(adir)
atlantis_stomach_content(adir,
        consumption,
        predator_map,
        prey_map)
}

\arguments{
    \item{adir}{The atlantis directory created by \link{atlantis_directory}}
    \item{consumption}{Survey data generated by \link{atlantis_fg_tracer} with \code{consumption = TRUE}}
    \item{predator_map}{Vector of functional groups to MFDB species, e.g. \code{c(FCD = 'COD')}}
    \item{prey_map}{Vector of functional groups to MFDB species, e.g. \code{c(FHE = 'HER')}, any other species will be discarded}
}

\value{
\code{atlantis_stomach_matrix} returns the content of the Atlantis diet matrix,
annotated with year/month.

\code{atlantis_stomach_content} expands the consumption data into a stomach
survey, dividing the total consumed biomass into proportions based on the diet
matrix.
}
