\name{R/atlantis_tracer.R}
\alias{atlantis_temperature}
\alias{atlantis_fg_tracer}
\alias{atlantis_tracer_add_lengthgroups}
\alias{atlantis_tracer_survey_select}

\title{Atlantis tracer parsers}
\description{
    Routines to read and make use of the tracer output files
}

\usage{
atlantis_temperature(adir,
        area_data)
atlantis_fg_tracer(adir,
        area_data,
        fg_group,
        consumption = FALSE)
atlantis_tracer_add_lengthgroups(
        tracer_data,
        length_group,
        sigma_per_cohort)
atlantis_tracer_survey_select(
        tracer_data,
        length_group,
        survey_suitability,
        survey_sigma)
}

\arguments{
    \item{adir}{The atlantis directory created by \link{atlantis_directory}}
    \item{area_data}{The area data returned by \link{atlantis_read_areas}}
    \item{fg_group}{One of the functional groups returned from \link{atlantis_functional_groups}}
    \item{consumption}{Include consumption data?}
    \item{tracer_data}{Output of \link{atlantis_fg_tracer} or \link{atlantis_tracer_add_lengthgroups}}
    \item{length_group}{Vector of c(min, min, min, ..., max) lengths}
    \item{sigma_per_cohort}{Vector of sigma per-cohort in age group}
    \item{survey_suitability}{Suitability, 0..1 value for each length group to scale abundance}
    \item{survey_sigma}{Error rate per-length-group}
}

\value{
\code{atlantis_temperature} returns a \code{data.frame} of temperatures
suitable for importing into MFDB.

\code{atlantis_fg_tracer} returns a \code{data.frame} containing total
abundance/length/weight information, and optionally consumption/day for
the given functional group. Can be imported into MFDB, but ideally a
survey is simulated first.

\code{atlantis_tracer_add_lengthgroups} break up abundance from \link{atlantis_fg_tracer}
into length groups, assuming a normal distribution around the
Atlantis-reported mean.

\code{atlantis_tracer_survey_select} samples from \link{atlantis_tracer_add_lengthgroups}
output based on the \code{survey_suitability} vector (possibly generated by
rgadget), adding noise based on \code{survey_sigma}. The result is a
\code{data.frame} that can be imported as a survey into MFDB.
}
